requirements.txt
Django==3.2.8
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Markdown==3.3.4
Pillow
drf-yasg
python-decouple
redis>=3.5
celery>=5.0
django-celery-beat
supervisor
psycopg2-binary
django-admin-rangefilter
django-jalali-date
mysqlclient
django-mptt
fcm-django==1.0.6
six==1.15.0
django-tinymce